refactored drinkdispenser assignment(inclass_7_1)

+++ inclass_7_1/src/com/company/CupHolder.java
@@ -9,7 +9,7 @@ public class CupHolder {
         this.quantity = quantity;
     }
 
-    public double getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
